Create UploadPart.go
To implement the UploadPart method in the S3Uploader struct, you need to use the UploadPart function from the s3manager package. This function is part of the multipart upload process, where a large file is divided into parts and each part is uploaded separately.

Here's how you might implement the UploadPart method:

In this implementation:

The UploadPart method takes a multipart.File which is the file part to be uploaded, the bucket name, the object key, the upload ID (which uniquely identifies the multipart upload session), and the part number (which identifies the part within the multipart upload).
It reads the file part into a buffer using ioutil.ReadAll.
It creates an UploadPartInput struct with the necessary information.
It calls the UploadPart method of the S3 service client to upload the part.
It returns the response from the UploadPart call or an error if something goes wrong.
Remember to replace "us-west-2" with the AWS region you're using and ensure that your AWS credentials are configured correctly. Also, this method assumes that the multipart upload has already been initiated, and you have a valid uploadID to use.